transform for target subschema separately
